requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
magic-filter==1.0.12
multidict==6.3.2
openpyxl==3.1.5
packaging==24.2
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
uvicorn==0.34.0
yarl==1.19.0
